requirements.txt
google-api-python-client==2.39.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
pytesseract==0.3.8
Pillow==9.2.0
openai>=1.0.0
python-dotenv>=0.21.0
bs4>=0.0.1
pydantic>=2.0.0
pytz>=2023.3
tiktoken>=0.5.0
colorama>=0.4.6